URL, App 이름 설정 및 RESTful

+++ django/crud_review/students/views.py
@@ -11,14 +11,23 @@ def index(request):
     return render(request, 'students/index.html', context)
 
 def new(request):
-    return render(request, 'students/new.html')
+    if request.method == 'POST':     # def create
+        name = request.POST.get('name')
+        age = request.POST.get('age')
 
-def create(request):
-    name = request.POST.get('name')
-    age = request.POST.get('age')
+        student = Student.objects.create(name = name, age = age)
+        return redirect('students:detail', student.pk) 
 
-    student = Student.objects.create(name = name, age = age)
-    return redirect(f'/students/{student.pk}/')
+    else:
+        # new page를 보여주면 됨 (def new)
+        return render(request, 'students/new.html')
+
+# def create(request):
+#     name = request.POST.get('name')
+#     age = request.POST.get('age')
+
+#     student = Student.objects.create(name = name, age = age)
+#     return redirect('students:detail', student.pk) 
 
 def detail(request, pk):
     student = Student.objects.get(pk=pk)
@@ -29,33 +38,49 @@ def detail(request, pk):
     return render(request, 'students/detail.html', context)
 
 def delete(request, pk):
-    student = Student.objects.get(pk=pk)
-    student.delete()
-    context = {
-        'student' : student,
-    }
-    return redirect('/students/')
+    if request.method = 'POST':
+        student = Student.objects.get(pk=pk)
+        student.delete()
+        context = {
+            'student' : student,
+        }
+        return redirect('students:index')
 
 def edit(request, pk):
     student = Student.objects.get(pk=pk)
-    context = {
-        'student' : student,
-    }
-    return render(request, 'students/edit.html', context)
 
-def update(request, pk):
+    if request.method == 'POST':        # def update
+        name = request.POST.get('name')
+        age = request.POST.get('age')
+        
+        student.name = name
+        student.age = age
+        student.save()
 
-    student = Student.objects.get(pk=pk)
+        context = {
+            'student' : student,
+        }
+        return redirect('students:detail', student.pk)
 
-    name = request.POST.get('name')
-    age = request.POST.get('age')
+    else:                               # def edit
+        context = {
+            'student' : student,
+        }
+        return render(request, 'students/edit.html', context)
+
+# def update(request, pk):
+
+#     student = Student.objects.get(pk=pk)
+
+#     name = request.POST.get('name')
+#     age = request.POST.get('age')
     
-    student.name = name
-    student.age = age
-    student.save()
+#     student.name = name
+#     student.age = age
+#     student.save()
 
-    context = {
-        'student' : student,
-    }
-    return redirect(f'/students/{student.pk}/')     # detail 페이지로 redirect
+#     context = {
+#         'student' : student,
+#     }
+#     return redirect('students:detail', student.pk)     # detail 페이지로 redirect
